// packages/api/src/lib/generated/plugin-host-modules.schema.generated.ts
export const pluginHostModuleRegistry = {
  "@codemirror/autocomplete": {
    entry: "@codemirror/autocomplete",
    moduleId: "codemirror-autocomplete",
    packageName: "@codemirror/autocomplete",
    platforms: ["web", "electron-renderer"],
    public: true,
    source: "config",
    specifier: "@codemirror/autocomplete",
    version: "6.18.6",
  },
// ...
} as const;